+++ NEWS.md
@@ -1,6 +1,7 @@
 # jsmodule 1.6.1
 ## Update
 - Add function to allow adjusting cutoff for a single independent variable and observing model's metrics in `rocModule`, `rocModule2`.
+- Add option to turn pairwise p value option on in case level of stratified group >= 3, in `tb1moduleUI`.
 
 # jsmodule 1.6.0 
 ## Update